requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
idna==3.3
loguru==0.6.0
pyTelegramBotAPI==4.11.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot-calendar==1.0.5
requests==2.28.1
six==1.16.0
urllib3==1.26.12
win32-setctime==1.1.0
peewee==3.16.2
